refactor: extract DWARF CFA and storage helpers
Move CFA expression lowering and storage-address discovery into the shared

dwarf_expr module so CFI and fast-parser callers no longer own expression

operation parsing details.
