+++ ghostscope-dwarf/src/dwarf_expr/mod.rs
@@ -1,6 +1,7 @@
 //! Shared DWARF expression parsing, scanning, and lowering utilities.
 
 pub(crate) mod call_site;
+pub(crate) mod cfa;
 pub(crate) mod const_eval;
 pub(crate) mod entry_value;
 pub(crate) mod errors;
@@ -8,5 +9,6 @@ pub(crate) mod lower;
 pub(crate) mod modes;
 pub(crate) mod ops;
 pub(crate) mod scan;
+pub(crate) mod storage;
 
 pub(crate) use lower::ExpressionEvaluator;
